requirements.txt
telethon==1.24.0
ffmpeg-python==0.2.0
